% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blance.R
\docType{package}
\name{blance}
\alias{blance}
\alias{blance-package}
\title{blance: Brian's personal convenience functions.}
\description{
Thank you to...
\itemize{
\item \href{https://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/}{Hilary Parker}, for getting me to explore package development in the first place.
\item \href{https://brendanrocks.com/blogging-with-rmarkdown-knitr-jekyll/}{Brendan Rocks}, for showing me how to blog in Rmarkdown without getting lost in the weeds (all blog-related functions are direct copies or revised versions of Brendan's own.)
}
}

