% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/culk.R
\name{load_orders}
\alias{load_orders}
\title{Load and Clean Stitch Orders}
\usage{
load_orders(f)
}
\arguments{
\item{f}{The path/name of the stitch export to open (must be .csv)}
}
\value{
A cleaned dataframe ready for analysis
}
\description{
Opens and performs basic cleaning on order data exported from Stitch Labs.
}

