% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmamp.R
\name{mamp_connect}
\alias{mamp_connect}
\title{Create R connection object to default local MAMP MySQL database}
\usage{
mamp_connect(dbname)
}
\arguments{
\item{dbname}{a string corresponding to a database name in MAMP.}
}
\value{
A MySQLConnection object is returned, and should be passed
   to a new variable when calling mampConnect.
}
\description{
This function returns a MySQLConnection object to the
named MAMP database (assumes default settings on OS X 10.11.5).
}
\details{
You must have MAMP installed and running to use this function.

Thank you to Andrew Knight for \href{http://apknightcode.blogspot.com/2012/09/first-post.html}{blogging about this}.
}

