% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blance.R
\docType{package}
\name{blance}
\alias{blance}
\alias{blance-package}
\title{blance: Brian's personal convenience functions.}
\description{
Thank you to...
\itemize{
\item \href{https://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/}{Hilary Parker}, for getting me to explore package development in the first place.
\item \href{https://brendanrocks.com/blogging-with-rmarkdown-knitr-jekyll/}{Brendan Rocks}, for a super easy intro to blogging in Rmarkdown (all blog-related functions are copied / adapted from Brendan's own)
\item \href{https://github.com/yihui/knitr-jekyll}{Yihui Xie}, for the knitr-jekyll repo which makes blogging in Rmarkdown possible for non-dev heads in the first place.
}
}

